This program demonstrates the use of the HomeStepper() function.
HomeStepper only needs to be called once to function. This program contains two functions, HomeStepper() and pin_ISR(). pin_ISR() is an interrupt service routine. HOMESTEPPER USES INTERRUPTS BE CAREFULL WITH THIS. Both functions must be placed within the main program for this functionality to work. Also note that the stepper library must be included and that the variable Home must a globally define variable.